cdk/requirements.txt
aws-cdk-lib==2.13.0
constructs>=10.0.0
aws_cdk.aws_batch_alpha